backend/api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import DetectionViewSet, ChatView

router = DefaultRouter()
router.register(r'detections', DetectionViewSet, basename='detection')

urlpatterns = [
    path('', include(router.urls)),
    path('chat/', ChatView.as_view(), name='chat'),
]