backend/requirements.txt
# Core Django framework
Django>=5.2.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter (if using PostgreSQL)
# sqlite3 is included with Python

# CORS support for React frontend
django-cors-headers>=4.0.0

# AI/ML Libraries
tensorflow>=2.13.0
tf-keras>=2.15.0,<3.0.0
numpy>=1.24.0
Pillow>=10.0.0  # PIL for image processing
opencv-python>=4.8.0  # cv2 for image operations
scikit-learn>=1.3.0  # For metrics and preprocessing

# YOLO and Computer Vision (existing)
ultralytics>=8.0.0  # YOLOv11 support
torch>=2.0.0  # PyTorch for YOLO
torchvision>=0.15.0

# Image processing
matplotlib>=3.7.0  # For plotting and visualization

# Additional utilities
python-dotenv>=1.0.0  # Environment variables
Pillow>=10.0.0  # Updated PIL

# RAG / LLM
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.5
sentence-transformers>=3.0.0
pypdf>=4.0.0
pinecone>=5.0.0
langchain-pinecone>=0.1.0
langchain-huggingface>=0.0.3